+++ dpdocter/src/main/java/com/dpdocter/request/RecordsSearchRequest.java
@@ -15,7 +15,7 @@ public class RecordsSearchRequest {
 
     private Boolean discarded = true;
 
-    private int page;
+    private long page;
 
     private int size;
 
@@ -77,11 +77,11 @@ public class RecordsSearchRequest {
 	this.discarded = discarded;
     }
 
-    public int getPage() {
+    public long getPage() {
 	return page;
     }
 
-    public void setPage(int page) {
+    public void setPage(long page) {
 	this.page = page;
     }
