+++ dpdocter/src/main/java/com/dpdocter/services/AdmitCardService.java
@@ -6,14 +6,13 @@ import org.bson.types.ObjectId;
 
 import com.dpdocter.request.AdmitCardRequest;
 import com.dpdocter.response.AdmitCardResponse;
-import com.sun.jersey.multipart.FormDataBodyPart;
 
 public interface AdmitCardService {
 
 	AdmitCardResponse getAdmitCard(String cardId);
 
 	List<AdmitCardResponse> getAdmitCards(String doctorId, String locationId, String hospitalId, String patientId,
-			int page, int size, long updatedTime, Boolean discarded);
+			long page, int size, long updatedTime, Boolean discarded);
 
 	AdmitCardResponse deleteAdmitCard(String cardId, String doctorId, String hospitalId, String locationId,
 			Boolean discarded);
