+++ dpdocter/src/main/java/com/dpdocter/services/ClinicContactUsService.java
@@ -9,10 +9,10 @@ public interface ClinicContactUsService {
 	
 	public String submitClinicContactUSInfo(ClinicContactUs clinicContactUs);
 
-	//public List<DoctorContactUs> getDoctorContactList(int page, int size);
+	//public List<DoctorContactUs> getDoctorContactList(long page, int size);
 
 	public ClinicContactUs updateClinicContactState(String contactId, DoctorContactStateType contactState);
 
-	public List<ClinicContactUs> getDoctorContactList(int page, int size, String searchTerm);
+	public List<ClinicContactUs> getDoctorContactList(long page, int size, String searchTerm);
 
 }
