+++ dpdocter/src/main/java/com/dpdocter/services/DentalImagingService.java
@@ -14,8 +14,6 @@ import com.dpdocter.beans.Hospital;
 import com.dpdocter.collections.DentalImagingInvoiceResponse;
 import com.dpdocter.request.DentalImagingLabDoctorRegistrationRequest;
 import com.dpdocter.request.DentalImagingReportsAddRequest;
-import com.dpdocter.request.DentalLabDoctorRegistrationRequest;
-import com.dpdocter.request.DoctorLabReportsAddRequest;
 import com.dpdocter.response.DentalImagingDataResponse;
 import com.dpdocter.response.DentalImagingLocationResponse;
 import com.dpdocter.response.DentalImagingLocationServiceAssociationLookupResponse;
@@ -24,25 +22,22 @@ import com.dpdocter.response.DentalImagingVisitAnalyticsResponse;
 import com.dpdocter.response.DoctorHospitalDentalImagingAssociationResponse;
 import com.dpdocter.response.PatientAnalyticResponse;
 import com.dpdocter.response.PatientDentalImagignVisitAnalyticsResponse;
-import com.dpdocter.response.ServiceLocationResponse;
-
-import common.util.web.Response;
 
 public interface DentalImagingService {
 
 	DentalImagingResponse addEditDentalImagingRequest(DentalImagingRequest request);
 
-	List<DentalDiagnosticService> getServices(String searchTerm, String type, int page, int size);
+	List<DentalDiagnosticService> getServices(String searchTerm, String type, long page, int size);
 
 	Boolean addEditDentalImagingLocationServiceAssociation(List<DentalImagingLocationServiceAssociation> request);
 
 	List<Hospital> getHospitalList(String doctorId, String hospitalId);
 
 	List<DentalImagingLocationResponse> getServiceLocations(List<String> dentalImagingServiceId, String hospitalId,
-			String searchTerm, int size, int page);
+			String searchTerm, int size, long page);
 
 	List<DentalImagingLocationServiceAssociationLookupResponse> getLocationAssociatedServices(String locationId,
-			String hospitalId, String searchTerm, String type, int page, int size, Boolean discarded);
+			String hospitalId, String searchTerm, String type, long page, int size, Boolean discarded);
 
 	DentalImagingReports addDentalImagingReportBase64(FileDetails fileDetails, DentalImagingReportsAddRequest request);
 
@@ -51,21 +46,21 @@ public interface DentalImagingService {
 	List<DoctorHospitalDentalImagingAssociation> getDoctorHospitalAssociation(String doctorId);
 
 	List<DentalImagingResponse> getRequests(String locationId, String hospitalId, String doctorId, Long from, Long to,
-			String searchTerm, int size, int page, String type);
+			String searchTerm, int size, long page, String type);
 
 	DentalImaging discardRequest(String id, boolean discarded);
 
 	DentalImagingReports discardReport(String id, boolean discarded);
 
 	List<DoctorHospitalDentalImagingAssociationResponse> getHospitalAssociatedDoctor(String hospitalId,
-			String searchTerm, int size, int page);
+			String searchTerm, int size, long page);
 
 	Boolean dentalLabDoctorRegistration(DentalImagingLabDoctorRegistrationRequest request);
 
 	//DentalImagingInvoice addEditInvoice(DentalImagingInvoice request);
 
 	Double getInvoiceAmount(String doctorId, String locationId, String hospitalId, String fromDate, String toDate,
-			String dentalImagingLocationId, String dentalImagingHospitalId, int page, int size);
+			String dentalImagingLocationId, String dentalImagingHospitalId, long page, int size);
 
 	DentalImagingInvoice discardInvoice(String id, boolean discarded);
 
@@ -73,7 +68,7 @@ public interface DentalImagingService {
 
 	List<DentalImagingInvoiceResponse> getInvoices(String doctorId, String locationId, String hospitalId,
 			String dentalImagingLocationId, String dentalImagingHospitalId, Long from, Long to, String searchTerm,
-			int size, int page, Boolean isPaid);
+			int size, long page, Boolean isPaid);
 
 	public DentalImagingInvoiceResponse getInvoice(String invoiceId);
 
@@ -82,7 +77,7 @@ public interface DentalImagingService {
 
 	List<DentalImagingReports> getReports(String doctorId, String locationId, String hospitalId,
 			String dentalImagingLocationId, String dentalImagingHospitalId, String patientId, Long from, Long to,
-			String searchTerm, int size, int page);
+			String searchTerm, int size, long page);
 
 	List<PatientAnalyticResponse> getPatientVisitAnalytics(Long fromDate, Long toDate, String dentalImagingLocationId,
 			String dentalImagingHospitalId, String searchType);
@@ -104,7 +99,7 @@ public interface DentalImagingService {
 
 	PatientDentalImagignVisitAnalyticsResponse getDetailedDoctorVisitAnalytics(Long fromDate, Long toDate,
 			String dentalImagingLocationId, String dentalImagingHospitalId, String doctorId, String searchType,
-			int page, int size);
+			long page, int size);
 
 	DentalImagingDataResponse getDentalImagingData();
