+++ dpdocter/src/main/java/com/dpdocter/services/DiagnosticTestOrderService.java
@@ -8,7 +8,7 @@ import com.dpdocter.response.LabSearchResponse;
 
 public interface DiagnosticTestOrderService {
 
-	List<LabSearchResponse> searchLabs(String city, String location, String latitude, String longitude, String searchTerm, List<String> testNames, int page, int size, Boolean havePackage);
+	List<LabSearchResponse> searchLabs(String city, String location, String latitude, String longitude, String searchTerm, List<String> testNames, long page, int size, Boolean havePackage);
 
 	List<DiagnosticTestSamplePickUpSlot> getDiagnosticTestSamplePickUpTimeSlots();
